requirements.txt
flask==1.0.2
grpcio==1.12.1
grpcio-tools==1.12.1
pytest==3.9.3
requests
protobuf==3.6.1
jsonpickle==1.2
numpy==1.17.2
Pillow==6.2.0